refactor: move NHL Edge types to api module
Move NHL Edge type declarations from /types/nhl-edge.ts to /lib/api/nhl-edge/types/
and update corresponding imports.
