+++ apps/sploosh-ai-hockey-analytics/components/features/games/card/game-card.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image'
-import { NHLEdgeGame } from '@/types/nhl-edge'
+import { NHLEdgeGame } from '@/lib/api/nhl-edge/types/nhl-edge'
 import { parseISO } from 'date-fns'
 import { formatInTimeZone } from 'date-fns-tz'
