+++ apps/sploosh-ai-hockey-analytics/lib/api/nhl-edge/services/play-by-play.ts
@@ -1,4 +1,4 @@
-import type { NHLEdgePlayByPlay } from '@/types/nhl-edge'
+import type { NHLEdgePlayByPlay } from '@/lib/api/nhl-edge/types/nhl-edge'
 
 export async function getPlayByPlay(gameId: string): Promise<NHLEdgePlayByPlay> {
     const response = await fetch(`/api/nhl/play-by-play?gameId=${gameId}`)
