+++ apps/sploosh-ai-hockey-analytics/lib/api/nhl-edge/services/scores.ts
@@ -1,4 +1,4 @@
-import type { NHLEdgeScheduleResponse } from '@/types/nhl-edge'
+import type { NHLEdgeScheduleResponse } from '@/lib/api/nhl-edge/types/nhl-edge'
 
 export async function getScores(date: string): Promise<NHLEdgeScheduleResponse> {
     const response = await fetch(`/api/nhl/scores?date=${date}`)
