+++ apps/sploosh-ai-hockey-analytics/lib/api/nhl-edge/types/nhl-edge.ts
