+++ apps/sploosh-ai-hockey-analytics/lib/utils/game-state.ts
@@ -1,4 +1,4 @@
-import type { NHLEdgeGame } from '@/types/nhl-edge'
+import type { NHLEdgeGame } from '@/lib/api/nhl-edge/types/nhl-edge'
 
 export function shouldEnableAutoRefresh(games: NHLEdgeGame[]): boolean {
     if (!games.length) return false;
